% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{BDD background}
\usage{
background(title, given, ...)
}
\arguments{
\item{title}{Background title}

\item{given}{Preconditions for scenario/feature.}

\item{...}{additional \code{and} arguments for initial context (provided in \code{list()})}
}
\value{
A BDD scenario
}
\description{
BDD backgrounds are listed \strong{before} the first \code{scenario()}, at the same
level of indentation.
}
\examples{
background(title = "Launching the application",
           given = "I have launched the application")
background(title = "Launching the application",
           given = "I have launched the application",
           and = "it contains movie review data from IMDB and Rotten Tomatoes")
background(title = "Launching the application",
           given = "I have launched the application",
           and = list("it contains movie review data from IMDB and Rotten Tomatoes",
                      "the data contains variables like 'Critics Score' and 'MPAA'",
                      "the data contains variables like 'Audience Score' and 'Genre'"))
}
