% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background_build}
\alias{background_build}
\title{BDD background (build)}
\usage{
background_build(title, given, and = NULL)
}
\arguments{
\item{title}{Background title}

\item{given}{Preconditions/initial context}

\item{and}{Additional 'and' arguments}
}
\value{
A BDD background
}
\description{
BDD backgrounds are listed \strong{before} the first \code{scenario()}, at the same
level of indentation.
}
\keyword{internal}
