% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_string.R
\name{check_string}
\alias{check_string}
\title{Check for string (rlang util)}
\usage{
check_string(x)
}
\arguments{
\item{x}{input argument}
}
\value{
\code{TRUE} is passes, alert + \code{FALSE} if fails
}
\description{
Check for string (rlang util)
}
\examples{
pickler:::check_string("test")
pickler:::check_string(1)
}
\keyword{internal}
