% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{collapse_transformer}
\alias{collapse_transformer}
\title{Collapse transformer (utility function)}
\usage{
collapse_transformer(regex = "[+]$", ...)
}
\arguments{
\item{regex}{regular expression}

\item{...}{arguments passed to \code{glue_collapse()}}
}
\value{
transformed string
}
\description{
A transformer which automatically collapses any glue block ending with \code{+}
}
\section{Source}{

This function comes from the \href{https://glue.tidyverse.org/articles/transformers.html}{glue website}
}

\examples{
glue::glue("{1:5+}\n{letters[1:5]+}",
           .transformer = collapse_transformer(sep = ", "))
glue::glue("{1:5+}\n{letters[1:5]+}",
           .transformer = collapse_transformer(sep = ", ", last = " and "))
x <- c("one", "two")
glue::glue("{x}: {1:5+}",
           .transformer = collapse_transformer(sep = ", "))
}
