% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lbl_col.R
\name{add_lbl_col}
\alias{add_lbl_col}
\title{Drop Row Names and Add a Label Column}
\usage{
add_lbl_col(element_name, parsed_data_list, col_name)
}
\arguments{
\item{element_name}{string element to be used as the label for new column.}

\item{parsed_data_list}{A list of \code{data.frame}s from which the \code{element_data}
data.frame is selected using \code{element_name}}

\item{col_name}{name of the new column to be added to the data.frame,
which will contain the value specified by \code{element_name}.}
}
\description{
The \code{add_lbl_col()} function is designed to modify a data frame
by removing its row names and appending a new column with a specific label.
}
