% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_describe_descriptions.R
\name{extract_describe_descriptions}
\alias{extract_describe_descriptions}
\title{Extract Describe Descriptions from Test Content}
\usage{
extract_describe_descriptions(test_content)
}
\arguments{
\item{test_content}{output from \code{read_test_file()}}
}
\value{
A character vector containing all the \code{descriptions} in the
\code{describe()} function calls extracted from the test content. If no
\code{describe()} functions, an empty vector is returned.

named string of describe descriptions
}
\description{
This function searches through the test contents for any \code{describe()}
functions, indicative of BDD tests in test frameworks like \code{testthat}.
}
\section{More Information}{

The function uses regular expressions to identify \code{describe()} functions.
}

