% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_expects.R
\name{extract_expects}
\alias{extract_expects}
\title{Extract and Combine Expectations from Parsed Test Blocks}
\usage{
extract_expects(test_code_list)
}
\arguments{
\item{test_code_list}{A list of parsed test code blocks. Each element of the
list is expected to be a \code{data.frame} with a structure identical to what is
returned by \code{utils::getParseData()}. These data frames represent parsed R
code, corresponding to test blocks in the test files.}
}
\value{
\code{extract_expects()} returns a \code{data.frame} that consolidates the
expectations extracted from each test file in the \verb{tests/testthat/} folder.
The \code{data.frame} contains merged results from the individual tests, with
additional columns for labeling and structuring the data. If no expectations
are found in any of the blocks, the function returns \code{NULL}.
}
\description{
\strong{extract}: 'verb as in physically remove, draw out'
}
