% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_its.R
\name{extract_it_descriptions}
\alias{extract_it_descriptions}
\title{Extract \code{description} from \code{it()}}
\usage{
extract_it_descriptions(test_content)
}
\arguments{
\item{test_content}{output from \code{read_test_file()}}
}
\value{
named string of \code{it()} \code{descriptions}.
}
\description{
Extract \code{description} from \code{it()}
}
