% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_thats.R
\name{extract_test_that_desc}
\alias{extract_test_that_desc}
\title{Extract \code{desc} from \code{test_that()}}
\usage{
extract_test_that_desc(test_content)
}
\arguments{
\item{test_content}{output from \code{read_test_file()}}
}
\value{
named string of \code{test_that()} descriptions (\code{desc})
}
\description{
Extract \code{desc} from \code{test_that()}
}
