% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature}
\alias{feature}
\title{BDD Feature}
\usage{
feature(As_a, I_want, So_that)
}
\arguments{
\item{As_a}{As a <user or stakeholder>}

\item{I_want}{I want <to perform an action>}

\item{So_that}{So that <I can reach business goal or deliver value>}
}
\value{
A feature
}
\description{
BDD Feature
}
\examples{
feature(As_a = "user",
    I_want = "to see the changes in the plot",
    So_that = "I can visualize the impact of my customizations")
}
