% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_build}
\alias{feature_build}
\title{BDD Feature (build)}
\usage{
feature_build(As_a, I_want, So_that)
}
\arguments{
\item{As_a}{"As a" <user/stakeholder>}

\item{I_want}{"I want" <to perform some action>}

\item{So_that}{"So that" <I can reach some business goal>}
}
\value{
A feature (based on [Gherkin](https://cucumber.io/docs/gherkin/) syntax)
}
\description{
BDD Feature (build)
}
\keyword{internal}
