% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_its.R
\name{get_its}
\alias{get_its}
\title{Read \code{it()} descriptions (contexts) from test file}
\usage{
get_its(test_file)
}
\arguments{
\item{test_file}{path to test file}
}
\value{
data.frame with \code{test_file}, \code{it} and \code{description}
}
\description{
Read \code{it()} descriptions (contexts) from test file
}
\examples{
get_its(test_file = system.file("tests", "testthat",
                                      "test-rev_string.R",
                                      package = "pickler"))
get_its(test_file = system.file("tests", "testthat",
                                      "test-logo.R",
                                      package = "pickler"))
}
