% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_thats.R
\name{get_test_thats}
\alias{get_test_thats}
\title{Read \code{test_that()} descriptions (context) from test file}
\usage{
get_test_thats(test_file)
}
\arguments{
\item{test_file}{path to test file}
}
\value{
data.frame with \code{test_file}, \code{test_that} and \code{desc}
}
\description{
Read \code{test_that()} descriptions (context) from test file
}
\examples{
get_test_thats(test_file = system.file("tests", "testthat",
                                      "test-rev_string.R",
                                      package = "pickler"))
get_test_thats(test_file = system.file("tests", "testthat",
                                      "test-logo.R",
                                      package = "pickler"))

}
