% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_expect.R
\name{match_expect}
\alias{match_expect}
\title{Match test expectations}
\usage{
match_expect(code_data, expectation_pattern = "^expect_")
}
\arguments{
\item{code_data}{A parse data data.frame representing parsed R code. The
structure should be what is returned by \code{utils::getParseData()}, containing
columns \code{token},  \code{text}, \code{line1}, \code{line2}, \code{terminal}, and \code{text}.}

\item{expectation_pattern}{The regular expression pattern that identifies
the start of the expectation functions defaults to \code{'^expect_'}.}
}
\value{
A data frame containing the extracted expectation functions and their
respective line numbers. This data.frame has columns \code{expectation},
\code{line1}, and \code{line2}. If no expectation functions are found, the function
returns \code{NULL}.
}
\description{
\strong{match}: 'verb (used with object): to equal or be equal to'
}
