% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_tests.R
\name{match_tests}
\alias{match_tests}
\title{Cull Test Calls from Tests}
\usage{
match_tests(test_data, context_pattern = "^context$")
}
\arguments{
\item{test_data}{A data frame representing a syntax tree of R code,
typically obtained from \code{utils::getParseData()}}

\item{context_pattern}{A regular expression as a string, used to match the
text content of tokens in the syntax tree that signify the start of a
test case. Defaults to \code{"^context$"} which is a placeholder for actual
test function names like \code{test_that}.}
}
\value{
A list of data frames, each corresponding to a nested test case.
The list is named using the first \code{SYMBOL_FUNCTION_CALL} token text in
each nested case. Each data frame in the list contains a subset of the
original syntax tree, corresponding to a specific test case.
}
\description{
Identifies and organizes test cases in R code.
}
\section{Technical details}{

This function takes a syntax tree as input, typically generated
by parsing R code, and organizes test cases based on the occurrence of
specified tokens. It is designed to work with test cases written using
common testing frameworks like \code{testthat}. The output is a list of nested
data frames, each representing a test case block.
}

