% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_test_file.R
\name{parse_test_file}
\alias{parse_test_file}
\title{Map the hierarchy structure of a testthat test file}
\usage{
parse_test_file(path)
}
\arguments{
\item{path}{character, path to file}
}
\value{
data.frame
}
\description{
Query a test file to see what expectations are defined in it.
}
