% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tests_dir.R
\name{parse_tests_dir}
\alias{parse_tests_dir}
\title{Extract the testthat calls from test files}
\usage{
parse_tests_dir(path = "tests/testthat")
}
\arguments{
\item{path}{character, path to file}
}
\value{
data.frame
}
\description{
\code{parse_tests_dir()} searches the tests in the \verb{tests/testthat/} folder
and creates a \code{data.frame} from the test calls.
}
