% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickle.R
\name{pickle}
\alias{pickle}
\title{Serialize object maintaining internal object references}
\usage{
pickle(object, connection = NULL, compress = TRUE, ...)
}
\arguments{
\item{object}{The object to 'pickle'.}

\item{connection}{an open connection, a string indicating the desired filename
or NULL. If NULL returns a raw vector.}

\item{compress}{if connection is a string, and compress is TRUE, opens a gzfile
to store the pickled object data to. If compress is FALSE no compression is applied.
Compress can also be one of bzip, xz, gzip to indicate the type of compression to apply.}

\item{...}{arguments passed to or from other methods.}
}
\value{
NULL unless connection = NULL, when the result is returned in a raw vector.
}
\description{
Recurses through the object hierarchy storing unique elements
    and the information necessary to reconstitute the object.
}
