% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickler_logo.R
\name{pickler_logo}
\alias{pickler_logo}
\title{Pickler logos}
\usage{
pickler_logo()
}
\arguments{
\item{size}{header size ("h1", "h2", "h3", "h4", "h5", "h6")}
}
\value{
HTML text logo
}
\description{
Pickler logos
}
