% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickler_logo.R
\name{pickler_logo2}
\alias{pickler_logo2}
\title{alternate pickler logo 2}
\usage{
pickler_logo2()
}
\arguments{
\item{size}{header size ("h1", "h2", "h3", "h4", "h5", "h6")}
}
\value{
text logo
}
\description{
logo with \verb{'}
}
\keyword{internal}
