% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_text.R
\name{process_text}
\alias{process_text}
\title{Process names and text in dataset}
\usage{
process_text(raw_data, fct = FALSE)
}
\arguments{
\item{fct}{convert factors to lowercase? If \code{TRUE}, factors are returned as
character}

\item{.data}{\code{data.frame}, \code{tibble}, or \code{data.table}}
}
\value{
processed dataset
}
\description{
Process names and text in dataset
}
\examples{
str(process_text(datasets::CO2))
str(process_text(datasets::CO2, fct = TRUE))
}
