% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_test_file.R
\name{read_test_file}
\alias{read_test_file}
\title{Read Test File Contents}
\usage{
read_test_file(test_file)
}
\arguments{
\item{test_file}{A string specifying the path to the test file to be read.
The function checks if the file exists before attempting to
read it.}
}
\value{
A data frame with two columns: \code{test_file} containing the name of
the test file and \code{test_contents} containing a string of the file's
contents. If the file does not exist, the function returns an
invisible empty result.
}
\description{
Reads the content of a specified test file and returns it in a structured
format.
}
\section{More Information}{

The function consolidates the contents of the test file into a single string,
which facilitates further processing, such as pattern matching. It is
particularly useful for analyzing test files in the context of automated
testing frameworks. \code{read_test_file()} is primarily used in conjunction with
the \code{extract_} functions.
}

\examples{
tests <- read_test_file(test_file = system.file("tests", "testthat",
                                                "test-rev_string.R",
                                                package = "pickler"))
tests
}
