% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{BDD scenario}
\usage{
scenario(Given, When, Then, ...)
}
\arguments{
\item{Given}{Some initial context}

\item{When}{An action occurs}

\item{Then}{The expected outcome or behavior}

\item{And}{additional 'And' arguments}
}
\value{
A BDD scenario
}
\description{
BDD scenario
}
\examples{
scenario(Given = "I have launched the application",
         When = "I interact with the sidebar controls",
         Then = "the graph should update with the selected options")
scenario(Given = "I have launched the application",
         And = "it contains movie review data from IMDB and Rotten Tomatoes",
         When = "I interact with the sidebar controls",
         Then = "the graph should update with the selected options")
scenario(Given = "I have launched the application",
         And = list("it contains movie review data from IMDB and Rotten Tomatoes",
                    "the data contains variables like 'Critics Score' and 'MPAA'"),
         When = "I interact with the sidebar controls",
         Then = "the graph should update with the selected options")
}
