% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario_build}
\alias{scenario_build}
\title{BDD scenario (build)}
\usage{
scenario_build(Given, When, Then, And = NULL)
}
\arguments{
\item{Given}{Some initial context}

\item{When}{An action occurs}

\item{Then}{The expected outcome or behavior}

\item{And}{additional 'And' arguments}
}
\value{
A BDD scenario
}
\description{
BDD scenario (build)
}
\keyword{internal}
