% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpickle.R
\name{unpickle}
\alias{unpickle}
\title{Deserialize pickle data recreating object structure including
    internal object references}
\usage{
unpickle(connection, ...)
}
\arguments{
\item{connection}{an open connection, a string indicating the name of source file
or a raw vector.}

\item{...}{arguments passed to or from other methods.}
}
\value{
NULL unless connection = NULL, when the result is returned in a raw vector.
}
\description{
Uses serialized pickle generated data to reconstitute an object
    of the same structure as was originally pickled. Internal shared object
    references are maintained in the new object.
}
