% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{BDD background}
\usage{
background(title, given, ...)
}
\arguments{
\item{title}{Background title}

\item{given}{Preconditions for scenario/feature.}

\item{...}{additional \code{and} arguments for initial context (provided in a
\code{list()})}
}
\value{
A BDD background (based on \href{https://cucumber.io/docs/gherkin/}{Gherkin} syntax)
}
\description{
BDD backgrounds provide context and pre-existing conditions for features and
scenarios.
}
\section{Technical details}{

Use \code{background()} to reduce repetitive information in scenarios or features.
}

\examples{
background(title = "Launching the application",
           given = "I have launched the application")
background(title = "Launching the application",
           given = "I have launched the application",
           and = "it contains movie review data from IMDB and Rotten Tomatoes")
background(title = "Launching the application",
           given = "I have launched the application",
           and = list("it contains movie review data from IMDB and Rotten Tomatoes",
                      "the data contains variables like 'Critics Score' and 'MPAA'",
                      "the data contains variables like 'Audience Score' and 'Genre'"))
}
\seealso{
Other {"BDD helpers"}: 
\code{\link{feature}()},
\code{\link{scenario}()}
}
\concept{{"BDD helpers"}}
