% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{Bundle BDD functions}
\usage{
bundle(...)
}
\arguments{
\item{...}{features, scenarios, background, etc.}
}
\value{
bundled features, scenarios, background, etc. to be passed into tests.
}
\description{
Bundle BDD functions
}
\examples{
bundle(
  feature(title = "Visualization",
          as_a = "user",
          i_want = "to see the changes in the plot",
          so_that = "I can visualize the impact of my customizations"),
  background(title = "Launching the application",
             given = "I have launched the application",
             and = "it contains movie review data from IMDB and Rotten Tomatoes"),
  scenario(title = "Viewing the Data Visualization",
           given = "I have launched the application",
           when = "I interact with the sidebar controls",
           then = "the graph should update with the selected options")
)
}
