% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature}
\alias{feature}
\title{BDD Feature}
\usage{
feature(title, as_a, i_want, so_that)
}
\arguments{
\item{title}{Feature title}

\item{as_a}{"As a \if{html}{\out{<user or stakeholder>}}"}

\item{i_want}{"I want \if{html}{\out{<to perform an action>}}"}

\item{so_that}{"So that \if{html}{\out{<I can reach business goal or deliver value>}}"}
}
\value{
A feature (based on \href{https://cucumber.io/docs/gherkin/}{Gherkin} syntax)
}
\description{
BDD Feature
}
\examples{
feature(title = "Visualization",
        as_a = "user",
        i_want = "to see the changes in the plot",
        so_that = "I can visualize the impact of my customizations")
}
