% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature_build}
\alias{feature_build}
\title{BDD Feature (build)}
\usage{
feature_build(title, as_a, i_want, so_that)
}
\arguments{
\item{title}{Feature title}

\item{as_a}{"As a" <user/stakeholder>}

\item{i_want}{"I want" \if{html}{\out{<to perform some action>}}}

\item{so_that}{"So that" \if{html}{\out{<I can reach some business goal>}}}
}
\value{
A feature (based on \href{https://cucumber.io/docs/gherkin/}{Gherkin} syntax)
}
\description{
BDD Feature (build)
}
\keyword{internal}
