% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_string_long.R
\name{pivot_string_long}
\alias{pivot_string_long}
\title{Pivot a string into a data.frame (long)}
\usage{
pivot_string_long(string, sep = "[^[:alnum:]]+")
}
\arguments{
\item{string}{a character vector.}

\item{sep}{separator pattern (set to \code{"[^[:alnum:]]+"})}
}
\value{
data.frame with \code{unique_items} and \code{term}
}
\description{
\strong{This is a function for testing pickler.}
}
\examples{
pivot_string_long("A large size in stockings is hard to sell.")
pivot_string_long(c("A large size in stockings is hard to sell.", "The first part of the plan needs changing." ))
}
