% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_string}
\alias{rev_string}
\title{Reverse strings (function for testing)}
\usage{
rev_string(string)
}
\arguments{
\item{string}{string}
}
\value{
reversed string
}
\description{
An example function for testing and writing BDD feature, scenarios, and
background (original code was adapted from \href{https://stackoverflow.com/questions/56428360/modify-the-object-without-using-return-in-r-function}{this StackOverflow post})
}
\examples{
## NOT RUN
# a <- "StackOverFlow"
# rev_string(a)
# a
}
