% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_string.R
\name{rev_string}
\alias{rev_string}
\title{Reverse strings (function for testing)}
\usage{
rev_string(string)
}
\arguments{
\item{string}{string}
}
\value{
reversed string
}
\description{
\strong{This is a function for testing pickler.}
}
\section{More info}{


The original code for this function is from \href{https://stackoverflow.com/questions/56428360/modify-the-object-without-using-return-in-r-function}{this StackOverflow post}. I've included it because of it uses \code{deparse(substitute())} and \code{assign()} (which makes for interesting tests).
}

\examples{
## NOT RUN
# a <- "StackOverFlow"
# rev_string(a)
# a
}
