% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_string.R
\name{rev_string}
\alias{rev_string}
\title{Reverse strings (test function)}
\usage{
rev_string(x)
}
\arguments{
\item{x}{string}
}
\value{
reversed string
}
\description{
From https://stackoverflow.com/questions/56428360/modify-the-object-without-using-return-in-r-function
}
\examples{
## Input
a <- "StackOverFlow"
rev_string(a)
a
}
