% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{BDD scenario}
\usage{
scenario(title, given, when, then, ...)
}
\arguments{
\item{title}{Scenario title}

\item{given}{Preconditions or initial context}

\item{when}{An action occurs}

\item{then}{The expected outcome or behavior}

\item{and}{additional 'and' arguments}
}
\value{
A BDD scenario
}
\description{
Scenarios illustrate a concrete example of a specific behavior.
}
\section{Keywords}{

Scenarios include 'Given', 'When', and 'Then' keywords (and sometimes
additional 'And' statements).
}

\examples{
scenario(title = "Viewing the Data Visualization",
         given = "I have launched the application",
         when = "I interact with the sidebar controls",
         then = "the graph should update with the selected options")
scenario(title = "Viewing the Data Visualization",
         given = "I have launched the application",
         and = "it contains movie review data from IMDB and Rotten Tomatoes",
         when = "I interact with the sidebar controls",
         then = "the graph should update with the selected options")
scenario(title = "Viewing the Data Visualization",
         given = "I have launched the application",
         and = list("it contains movie review data from IMDB and Rotten Tomatoes",
                    "the data contains variables like 'Critics Score' and 'MPAA'"),
         when = "I interact with the sidebar controls",
         then = "the graph should update with the selected options")
}
