% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario}
\alias{scenario}
\title{BDD scenario}
\usage{
scenario(title, given, when, then, and = NULL)
}
\arguments{
\item{title}{Scenario title}

\item{given}{Preconditions or initial context}

\item{when}{An action occurs}

\item{then}{The expected outcome or behavior}

\item{and}{additional 'and' arguments}
}
\value{
A scenario (based on \href{https://cucumber.io/docs/gherkin/}{Gherkin} syntax)
}
\description{
Scenarios illustrate a concrete example of a specific behavior.
}
\section{Keywords}{

Scenarios include 'Given', 'When', and 'Then' keywords (and sometimes
additional 'And' statements).
}

\examples{
scenario(
    title = "Split column with a default pattern",
    given = "a dataframe with a specified column [name]",
    when = "I split the [name] column using the default [pattern]",
    then = "Then the [name] column should be split into multiple columns",
    and = "the new columns should be named with the default prefix [cols]"
  )
}
\seealso{
\code{\link[=feature]{feature()}}

Other {"BDD helpers"}: 
\code{\link{background}()},
\code{\link{feature}()}
}
\concept{{"BDD helpers"}}
