% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario.R
\name{scenario_build}
\alias{scenario_build}
\title{BDD scenario (build)}
\usage{
scenario_build(title, given, when, then, and = NULL)
}
\arguments{
\item{title}{Scenario title}

\item{given}{Some initial context}

\item{when}{An action occurs}

\item{then}{The expected outcome or behavior}

\item{and}{additional 'and' arguments}
}
\value{
A BDD scenario
}
\description{
BDD scenario (build)
}
\keyword{internal}
