% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_table.R
\name{with_table}
\alias{with_table}
\title{Add Gherkin-style table}
\usage{
with_table(tbl)
}
\arguments{
\item{tbl}{data.frame or tibble}
}
\value{
a table in \code{knitr::kable(format = "pipe", align = "l")}
}
\description{
\href{https://cucumber.io/docs/gherkin/reference/#scenario-outline}{Tables in Gherkin} are essentially markdown tables, so this is a wrapper for \code{knitr::kable(format = "pipe")}
}
\examples{
# with_table(data.frame(value = c(29L, 91L, 39L, 28L, 12L),
#                       name = c("John", "John, Jacob",
#                                "John, Jacob, Jingleheimer",
#                                "Jingleheimer, Schmidt",
#                                "JJJ, Schmidt")))
}
